added docs for chars, fixed some floating point tests

+++ rust/src/helpers.rs
@@ -10,7 +10,7 @@
 ///
 /// assert_eq!(print_type_of(&String::from("test")), "alloc::string::String");
 /// assert_eq!(print_type_of(&5), "i32");
-/// assert_eq!(print_type_of(&(5 as u8)), "u8");
+/// assert_eq!(print_type_of(&(5_u8)), "u8");
 /// assert_eq!(print_type_of(&[1]), "[i32; 1]");
 /// assert_eq!(print_type_of(&vec![""]), "alloc::vec::Vec<&str>");
 /// ```
@@ -22,7 +22,7 @@ pub fn print_type_of<T>(_: &T) -> String {
 fn print_type_of_works() {
 	assert_eq!(print_type_of(&String::from("test")), "alloc::string::String");
 	assert_eq!(print_type_of(&5), "i32");
-	assert_eq!(print_type_of(&(5 as u8)), "u8");
+	assert_eq!(print_type_of(&(5_u8)), "u8");
 	assert_eq!(print_type_of(&[1]), "[i32; 1]");
 	assert_eq!(print_type_of(&vec![""]), "alloc::vec::Vec<&str>");
 }
